refactor: refactoring folder structure
