--- packages/mobile/src/components/molecules/SettingHeader/style.module.scss
@@ -1,35 +1,0 @@
-@import "@shared/styles/color.scss";
-@import "@shared/styles/variables.scss";
-@import "@shared/styles/mixin.scss";
-
-.header {
-  position: fixed;
-  top: 0;
-  right: 0;
-  left: 0;
-  z-index: 100;
-  display: flex;
-  justify-content: space-between;
-  height: 70px;
-  background-color: $white;
-}
-
-.back-btn {
-  position: fixed;
-  display: block;
-  padding: 26px 22px 25px 21px;
-  line-height: 0;
-  background-color: $white;
-}
-
-.title {
-  margin: 0 auto;
-  padding: 24px 0 13px;
-  letter-spacing: 0.02em;
-  font-weight: 500;
-  font-size: 18px;
-  line-height: 23px;
-  color: $black-100;
-
-  @include ellipsis;
-}
