+++ packages/mobile/src/page/Setting/Main/style.module.scss
