+++ packages/mobile/src/page/Setting/SettingTemplate/style.module.scss
